+++ src/annotations_query_ctrl.ts
@@ -23,13 +23,13 @@ export class KustoDBAnnotationsQueryCtrl {
 
     return this.datasource
       .getDatabases()
-      .then(list => {
+      .then((list) => {
         this.databases = list;
         if (list.length > 0 && !this.annotation.database) {
           this.annotation.database = list[0].value;
         }
         return this.databases;
       })
-      .catch(err => {});
+      .catch((err) => {});
   }
 }