+++ src/module.ts
@@ -1,13 +1,14 @@
-import { DataSourcePlugin } from '@grafana/data';
+import { DataSourceApi, DataSourcePlugin } from '@grafana/data';
 import { AdxDataSource } from './datasource';
 import { QueryEditor } from './QueryEditor';
-import { KustoDBAnnotationsQueryCtrl } from './annotations_query_ctrl';
 import { AdxDataSourceOptions, AdxDataSourceSecureOptions, KustoQuery } from './types';
 import ConfigEditor from 'components/ConfigEditor';
 
-export const plugin = new DataSourcePlugin<AdxDataSource, KustoQuery, AdxDataSourceOptions, AdxDataSourceSecureOptions>(
-  AdxDataSource
-)
+export const plugin = new DataSourcePlugin<
+  DataSourceApi<KustoQuery, AdxDataSourceOptions>,
+  KustoQuery,
+  AdxDataSourceOptions,
+  AdxDataSourceSecureOptions
+>(AdxDataSource as any)
   .setConfigEditor(ConfigEditor)
-  .setAnnotationQueryCtrl(KustoDBAnnotationsQueryCtrl)
   .setQueryEditor(QueryEditor);