+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Change Log
 
-## [next]
+## [3.6.1]
+
+- Reverted change made in 3.6.0 and reuse the previous code editor until we fix the related issues.
+
+## [3.6.0]
+
+- Replaced custom query editor with @grafana/ui common editor, with support for Kusto.
+
+## [3.5.1]
+
+- Bugfix: Fixed issue where HTTP timeout setting was not being applied
 - Bugfix: Fixed issue when typing vs copy/pasting client secret in configuration
+- Bugfix: Fixed issue where annotation queries were not being displayed
 
 ## [3.5.0]